% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_crops.R
\name{add_crops}
\alias{add_crops}
\title{add_crops}
\usage{
add_crops(df, mapping = "mapping.xlsx")
}
\arguments{
\item{df}{dataframe with "j" or "c" column}

\item{mapping}{Mapping file which contains crops/activities}
}
\value{
Cleaned scenario name
}
\description{
add_crops
}
\examples{
\dontrun{
add_crops()
}
}
\author{
Abhijeet Mishra
}
