% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_regions.R
\name{add_regions}
\alias{add_regions}
\title{add_regions}
\usage{
add_regions(df)
}
\arguments{
\item{df}{dataframe with "cty" column}
}
\value{
Cleaned scenario name
}
\description{
add_regions
}
\examples{
\dontrun{
add_regions()
}
}
\author{
Abhijeet Mishra
}
