% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_filename.R
\name{clean_filename}
\alias{clean_filename}
\title{clean_filename}
\usage{
clean_filename(gdx)
}
\arguments{
\item{gdx}{GDX file for IMPACT run}
}
\value{
Cleaned scenario name
}
\description{
clean_filename
}
\examples{
\dontrun{
clean_filename()
}
}
\author{
Abhijeet Mishra
}
