% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_identifier_columns.R
\name{create_identifier_columns}
\alias{create_identifier_columns}
\title{create_identifier_columns}
\usage{
create_identifier_columns(
  df,
  col = "scenario",
  into = c("SSP", "GCM", "RCP", "CO2"),
  sep = "-",
  remove = TRUE
)
}
\arguments{
\item{df}{Dataframe where identifier columns need to be created}

\item{col}{Columns which are "ID" and need to be separated}

\item{into}{String vector with names of new columns}

\item{sep}{Separator where "col" columns should be snipped for new columns.}

\item{remove}{If "col" vector should be kicked out when making new columns.}
}
\value{
Cleaned scenario name
}
\description{
create_identifier_columns
}
\examples{
\dontrun{
create_identifier_columns()
}
}
\author{
Abhijeet Mishra
}
