% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMapping.R
\name{getMapping}
\alias{getMapping}
\title{getMapping}
\usage{
getMapping(type = "region", file = "mapping.xlsx")
}
\arguments{
\item{type}{"region" or "crops". Region returns six CGIAR regions + developed
or developing countries + world. Crops returns IMPACT mapping for crops in
their more user friendly name.}

\item{file}{mapping file name}
}
\value{
mapping according to the type selected
}
\description{
getMapping
}
\examples{
\dontrun{
getMapping()
}
}
\author{
Abhijeet Mishra
}
