% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group3.R
\name{group3}
\alias{group3}
\title{Group 3 data processing (Area, yields, production, prices, net trade, demand)}
\usage{
group3(gdx, indicator = "population")
}
\arguments{
\item{gdx}{GDX of an IMPACT run}

\item{indicator}{Which indicator to return in aggregated format. Defaults to
"population". Available settings are "area", "production", "yield", "prices",
"trade", "demand", "population" and "perCapDemand".}
}
\value{
dataframe results for Area, yields, production, prices, net trade,
demand
}
\description{
Group 3 data processing (Area, yields, production, prices, net trade, demand)
}
\examples{
\dontrun{
group3()
}
}
\author{
Abhijeet Mishra
}
