% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group4.R
\name{group4}
\alias{group4}
\title{Results for per capital kcal availability}
\usage{
group4(gdx)
}
\arguments{
\item{gdx}{GDX of an IMPACT run}
}
\value{
dataframe results for per capital kcal availability
}
\description{
Results for per capital kcal availability
}
\examples{
\dontrun{
group4()
}
}
\author{
Abhijeet Mishra
}
