% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group5.R
\name{group5}
\alias{group5}
\title{Results for population at risk of hunger}
\usage{
group5(gdx)
}
\arguments{
\item{gdx}{GDX of an IMPACT run}
}
\value{
Results for population at risk of hunger
}
\description{
Results for population at risk of hunger
}
\examples{
\dontrun{
group5()
}
}
\author{
Abhijeet Mishra
}
