% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_ggplot.R
\name{save_ggplot}
\alias{save_ggplot}
\title{save_ggplot}
\usage{
save_ggplot(
  p,
  dest = tempdir(),
  plot_name = "dummy",
  ext = "png",
  units = "in",
  width = 7,
  height = 7,
  dpi = 300
)
}
\arguments{
\item{p}{"plot" to be saved. Has to be a ggplot object}

\item{dest}{Destination directory of the plot}

\item{plot_name}{Name of the plot to be exported}

\item{ext}{Extension for saving the plot. e.g "png" or "eps" etc.}

\item{units, width, height}{Plot size in units ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types}
}
\value{
NULL. The function only saves a ggplot object
}
\description{
save_ggplot
}
\examples{
\dontrun{
save_ggplot()
}
}
