% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group1.R
\name{group1}
\alias{group1}
\title{Results for population and GDP}
\usage{
group1(gdx)
}
\arguments{
\item{gdx}{GDX of an IMPACT run}
}
\value{
dataframe results for population and GDP
}
\description{
Results for population and GDP
}
\examples{
\dontrun{
group1()
}
}
